Move one test from run-make-fulldeps to ui
