--- src/test/run-make-fulldeps/extern-prelude/basic.rs
@@ -1,6 +1,0 @@
-#![feature(extern_prelude)]
-
-fn main() {
-    let s = ep_lib::S; // It works
-    s.external();
-}
