--- src/test/run-make-fulldeps/extern-prelude/relative-only.rs
@@ -1,9 +1,0 @@
-// Extern prelude names are not available by absolute paths
-
-#![feature(extern_prelude)]
-
-use ep_lib::S;
-
-fn main() {
-    let s = ::ep_lib::S;
-}
