--- src/test/run-make-fulldeps/extern-prelude/shadow-mod.rs
@@ -1,14 +1,0 @@
-// Local module shadows `ep_lib` from extern prelude
-
-mod ep_lib {
-    pub struct S;
-
-    impl S {
-        pub fn internal(&self) {}
-    }
-}
-
-fn main() {
-    let s = ep_lib::S;
-    s.internal(); // OK
-}
