--- src/test/run-make-fulldeps/extern-prelude/shadow-prelude.rs
@@ -1,7 +1,0 @@
-// Extern prelude shadows standard library prelude
-
-#![feature(extern_prelude)]
-
-fn main() {
-    let x = Vec::new(0f32, ()); // OK
-}
