+++ src/test/ui/auxiliary/extern-prelude-vec.rs
@@ -1,3 +1,3 @@
-#![crate_type = "rlib"]
+#![crate_name = "Vec"]
 
 pub fn new(arg1: f32, arg2: ()) {}
