+++ src/test/ui/auxiliary/extern-prelude.rs
@@ -1,5 +1,3 @@
-#![crate_type = "rlib"]
-
 pub struct S;
 
 impl S {
