+++ src/test/ui/extern-prelude-fail.rs
@@ -1,0 +1,9 @@
+// compile-flags:--extern extern_prelude
+// aux-build:extern-prelude.rs
+
+// Extern prelude names are not available by absolute paths
+
+fn main() {
+    use extern_prelude::S; //~ ERROR unresolved import `extern_prelude`
+    let s = ::extern_prelude::S; //~ ERROR failed to resolve
+}
